LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/window-onerror-parse-error-expected.txt

CONSOLE MESSAGE: line 34: SyntaxError: Unexpected identifier 'script'
CONSOLE MESSAGE: line 1321: Error: Tried to create a test with file_is_test
CONSOLE MESSAGE: line 1321: Error: Tried to create a test with file_is_test
CONSOLE MESSAGE: line 1321: Error: Tried to create a test with file_is_test

PASS window.onerror: parse errors 
